refactor(orchestrator): split conductor module up
