+++ veecle-orchestrator/src/api.rs
@@ -17,8 +17,8 @@ use tracing::Instrument;
 use veecle_net_utils::{AsyncSocketStream, UnresolvedMultiSocketAddress};
 use veecle_orchestrator_protocol::{Info, InstanceId, Request, Response};
 
-use crate::conductor::Conductor;
 use crate::distributor::Distributor;
+use crate::runtime::Conductor;
 
 type Responder = Box<
     dyn FnOnce(
