+++ veecle-orchestrator/src/main.rs
@@ -10,15 +10,15 @@ use tracing_subscriber::filter::LevelFilter;
 use tracing_subscriber::layer::SubscriberExt;
 use veecle_net_utils::{UnresolvedMultiSocketAddress, UnresolvedSocketAddress};
 
-use self::conductor::Conductor;
 use self::distributor::Distributor;
+use self::runtime::Conductor;
 use self::telemetry::Exporter;
 
 mod api;
-mod conductor;
 mod distributor;
 mod external;
 mod eyre_tracing_error;
+mod runtime;
 mod telemetry;
 
 #[derive(Parser)]
